:truck: refactor(api/endpoint): Move reset dev endpoint.
