+++ imports/api/endpoint/_register/available.ts
@@ -1,4 +1,7 @@
 // eslint-disable-next-line import/no-unassigned-import
+import '../_dev/reset';
+
+// eslint-disable-next-line import/no-unassigned-import
 import '../allergies/changeColor';
 
 // eslint-disable-next-line import/no-unassigned-import
@@ -90,6 +93,3 @@ import '../uploads/delete';
 import '../uploads/rename';
 // eslint-disable-next-line import/no-unassigned-import
 import '../uploads/restore';
-
-// eslint-disable-next-line import/no-unassigned-import
-import '../testing/reset';
